add generic apiview class
