+++ api/urls.py
@@ -1,12 +1,13 @@
 from django.urls import path
 from . import views
 
+from .views import FlatListView, FlatRetrieveView, FlatCreateView, FlatRetrieveUpdateView, FlatDeleteView
+
 urlpatterns = [
     path('', views.index, name="index"),
-    path('api/flat-list/', views.flat_list, name='flat-list'),
-    path('api/flat-info/<str:pk>', views.flat_info, name='flat-info'),
-    path('api/create-flat', views.flat_create, name='create-flat'),
-    path('api/update-flat/<str:pk>', views.flat_update, name='update-flat'),
-    path('api/delete-flat/<str:pk>', views.flat_delete, name='delete-flat'),
+    path('api/', FlatListView.as_view(), name='flat_list'),
+    path('api/<int:pk>', FlatRetrieveUpdateView.as_view(), name='flat_info'),
+    path('api/create', FlatCreateView.as_view(), name='create_flat'),
+    path('api/delete/<int:pk>', FlatDeleteView.as_view(), name='delete_flat'),
 ]
